refactor(aio): move generation normalizer ownership
