+++ tests/test_python_backend_analyzer.py
@@ -683,9 +683,9 @@ ignored/
 
         self.assertEqual(analyzer.render_json(report), expected_text)
         self.assertEqual(report["schema_version"], 2)
-        self.assertEqual(report["inventory"]["module_count"], 67)
-        self.assertEqual(len(report["registry"]["shipped_python_modules"]), 67)
-        self.assertEqual(len(report["registry"]["runtime_import_closure"]), 67)
+        self.assertEqual(report["inventory"]["module_count"], 68)
+        self.assertEqual(len(report["registry"]["shipped_python_modules"]), 68)
+        self.assertEqual(len(report["registry"]["runtime_import_closure"]), 68)
         self.assertEqual(report["registry"]["missing_internal_imports"], [])
         self.assertEqual(report["registry"]["unreachable_shipped_python_modules"], [])
         self.assertTrue(
@@ -697,6 +697,7 @@ ignored/
                 "easyuse_anima/aio/generation_features.py",
                 "easyuse_anima/aio/generation_detailer.py",
                 "easyuse_anima/aio/generation_output.py",
+                "easyuse_anima/aio/generation_normalization.py",
                 "easyuse_anima/aio/generation_migrations.py",
                 "easyuse_anima/aio/generation_settings.py",
                 "easyuse_anima/common/__init__.py",
